Added support for limiting the mime type of data in MZRemoteData.
TagChimp will return an error html file instead of a 404 error status when an
image is not found. To handle this I have added support for limiting the
accepted mime type of data in MZRemoteData. Every place that expects to
download an image using MZRemoteData I have then set the accepted mime-type
to image/*.