Changed time zone to GMT-1100. Fixes #32.
The release date written to the file is a full date in the UTC time zone.
And different programs will display this date differently. iTunes uses the
local time zone while Plex displays the date in UTC time zone. When converting
from a day like 2013-11-27 to a full date the time part is set to midnight in
the default time zone and by setting that time zone to GMT-1100 we maximize the
number of other time zones that will be on the same day.