Fix one more dark mode issue
The `MZTextView` was not always setting the text color to
`NSColor.textColor`. This seems to be a bug fixed in later versions of
Xcode but since we still compile on High-Sierra we needed to find a fix.

The problem seemed to be that when the initial `textColor` and
`insertionPointColor` is `nil` (i.e. default color) the text color will
be wrong until the first time a non-empty string is shown from a loaded
value of any file or a search result. After the first color change the
value is correct.