Support TV app and missing iTunes
On a fresh installation of Catalina iTunes does not exist and so the
AppleScript plugins that deal with refreshing iTunes fail in that
situation. And even when iTunes is found on Catalina MetaZ should
instead talk to the TV app.

Catalina also included a new TV app which contains the Movie and TV Show
library previously in iTunes. This commit adds plugins, services and
clipboard support for the TV app on Catalina and all functionality that
MetaZ provided for iTunes is now supported for the TV app.