Added a URLSession cache for remote data downloading
When working with a tv series, I found that the loading of the
poster information for each tv episode re-loads ALL of the
posters for the series for each episode, even if the poster
data was the same for the previously selected episode.

Adding a URL cache to the URLSession to avoid re-downloading
(ive fixed the cache size to 128 meg) makes working with an
entire TV series a MUCH smoother experience.

For example, for a 'popular' tv series, seeing the first poster
for a selected search result might take 10-20 seconds
(as it was downloading 20-30 posters in the background before
displaying). For each episode in the series, it would re-download
the same URL data for the posters.  Adding the cache here means
it is only downloaded once.